fix(compliance-page): add benchmark overview detail, fix accounts name
