+++ src/locales/en-US/messages.po
@@ -73,7 +73,7 @@ msgstr "{0, plural, one {One error reported} two {Two errors reported} other {# 
 msgid "{0} ago"
 msgstr "{0} ago"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:146
+#: src/pages/panel/compliance/ComplianceDetail.tsx:152
 msgid "{0} categories"
 msgstr "{0} categories"
 
@@ -116,7 +116,7 @@ msgstr "{0} Regions"
 msgid "{0} regions, {1} countries"
 msgstr "{0} regions, {1} countries"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:166
+#: src/pages/panel/compliance/ComplianceDetail.tsx:172
 msgid "{0}_{accountName}_{1}.json"
 msgstr "{0}_{accountName}_{1}.json"
 
@@ -319,10 +319,6 @@ msgstr "Access to your account is broken"
 msgid "Account"
 msgstr "Account"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:180
-msgid "Account name"
-msgstr "Account name"
-
 #: src/pages/panel/inventory/getColumns.tsx:97
 msgid "Account usage"
 msgstr "Account usage"
@@ -428,7 +424,8 @@ msgstr "All"
 msgid "All {name}"
 msgstr "All {name}"
 
-#: src/pages/panel/compliance/CompliancePage.tsx:93
+#: src/pages/panel/compliance/ComplianceDetail.tsx:35
+#: src/pages/panel/compliance/ComplianceSummary.tsx:87
 msgid "All accounts"
 msgstr "All accounts"
 
@@ -627,6 +624,7 @@ msgstr "Benchmark"
 msgid "Benchmark: {0}"
 msgstr "Benchmark: {0}"
 
+#: src/pages/panel/compliance/CompliancePage.tsx:22
 #: src/pages/panel/resource-detail/utils/ResourceDetailChangeLogSelectedHistoryAccordion.tsx:54
 #: src/pages/panel/shared/utils/chartToShow.tsx:16
 #: src/pages/panel/user-settings-notifications/UserSettingsNotificationsPage.tsx:27
@@ -659,11 +657,11 @@ msgstr "Book a call"
 msgid "Business"
 msgstr "Business"
 
-#: src/pages/panel/compliance/CompliancePage.tsx:89
+#: src/pages/panel/compliance/ComplianceSummary.tsx:83
 msgid "By account"
 msgstr "By account"
 
-#: src/pages/panel/compliance/CompliancePage.tsx:77
+#: src/pages/panel/compliance/ComplianceSummary.tsx:71
 msgid "By framework"
 msgstr "By framework"
 
@@ -872,7 +870,7 @@ msgstr "Clouds"
 msgid "Community support"
 msgstr "Community support"
 
-#: src/pages/panel/compliance/CompliancePage.tsx:72
+#: src/pages/panel/compliance/ComplianceSummary.tsx:66
 #: src/pages/panel/security/OverallCard.tsx:76
 #: src/shared/layouts/panel-layout/menuList.tsx:30
 msgid "Compliance"
@@ -882,7 +880,7 @@ msgstr "Compliance"
 msgid "Compliance Benchmarks"
 msgstr "Compliance Benchmarks"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:183
+#: src/pages/panel/compliance/ComplianceDetail.tsx:191
 msgid "Compliance posture"
 msgstr "Compliance posture"
 
@@ -910,7 +908,7 @@ msgstr "Compute Changes"
 msgid "Compute changes in the last 30 days."
 msgstr "Compute changes in the last 30 days."
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:158
+#: src/pages/panel/compliance/ComplianceDetail.tsx:164
 msgid "Condense all"
 msgstr "Condense all"
 
@@ -1323,7 +1321,7 @@ msgstr "Everything in Free, and"
 msgid "Example searches"
 msgstr "Example searches"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:161
+#: src/pages/panel/compliance/ComplianceDetail.tsx:167
 msgid "Expand all"
 msgstr "Expand all"
 
@@ -2123,7 +2121,7 @@ msgstr "Owner"
 msgid "Passed"
 msgstr "Passed"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:186
+#: src/pages/panel/compliance/ComplianceDetail.tsx:194
 #: src/pages/panel/compliance/CompliancePassedChecksCard.tsx:26
 msgid "Passed checks"
 msgstr "Passed checks"
@@ -2615,7 +2613,7 @@ msgstr "Severities"
 msgid "Severity"
 msgstr "Severity"
 
-#: src/pages/panel/compliance/ComplianceDetail.tsx:153
+#: src/pages/panel/compliance/ComplianceDetail.tsx:159
 msgid "Show empty categories"
 msgstr "Show empty categories"
