+++ src/pages/panel/shared/utils/index.ts
@@ -5,6 +5,7 @@ export { FrameworkIcon } from './FrameworkIcon'
 export { getColorBySeverity } from './getColorBySeverity'
 export { getIconBySeverity } from './getIconBySeverity'
 export { getMaterialColorBySeverity } from './getMaterialColorBySeverity'
+export { HeatmapCard } from './HeatmapCard'
 export { isValidProp } from './isValidProp'
 export { navigateSubtitleQuery } from './navigateSubtitleQuery'
 export { nodeChangeToString, renderNodeChangeCell } from './nodeChangeCell'
